refactor: replace enums with Literal for TertiaryStructurePredictionMethod and AminoAcidRepresentation
