+++ src/graph_construction/graphs.py
@@ -16,8 +16,6 @@ from src.config.enum import (
     ExecutionMode,
     ESM2ModelForContactMap,
     ESM2Representation,
-    TertiaryStructurePredictionMethod,
-    AminoAcidRepresentation,
     EdgeConstructionFunctions,
     DistanceFunction
 )
@@ -30,9 +28,9 @@ class ConstructGraphDTO(DTO):
     execution_mode: ExecutionMode
     validation_mode: Optional[ValidationMode]
     randomness_percentage: Optional[PositiveFloat]
-    tertiary_structure_method: Optional[TertiaryStructurePredictionMethod]
+    tertiary_structure_method: Optional[str]
     pdb_path: Optional[Path]
-    amino_acid_representation: AminoAcidRepresentation
+    amino_acid_representation: str
     non_pdb_bound_sequences_file: Path
     edge_construction_functions: List[EdgeConstructionFunctions]
     distance_function: Optional[DistanceFunction]
